Functional card list search and display
Still work in progress, but the app is functional for filtering and searching for cards, with different types of HTML elements for different filters.
individual card modal is still to be developped further and that would most likely conclude the first step of the app which would make it a card search engine.
Next two steps would now be :
-to allow users to create lists of owned cards as well as desired cards.
-to allow users to create and edit decklists with desired cards, save them into localStorage but additionally export them into .txt files in a conventional format readable by Pokemon TCG Live or other applications related to Pokemon TCG.

A hypothetical final step would be to allow users to simulate a game using the deck of their choice, in a "table-top" setting, i.e users would manually and freely manipulate the cards in the environment as opposed to game rulings, restrictions and functionalities being fully coded.