Last commit corrected
A few errors were detected in the last commit version and are now fixed.
Added the option to filter by legality in standard format as an individual checkbox.
Modal behavior adjusted although still not entirely satisfying. More details to come for modals.
Step 1 of the app (card search engine) is almost finished.